+++ engine/webnodehonojs/src/index.js
@@ -22,7 +22,6 @@ module.exports = async (...args) => {
   return new Promise(async (resolve, reject) => {
     const [params, obj] = args;
     const [library, sys, cosetting] = obj;
-    // const { reaction } = library.utils;
     try {
       let webserver = await require("./webserver")(params, obj);
       let reaction = await require("./reaction")(params, obj);